% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DEEButil-package.R
\docType{package}
\name{DEEButil-package}
\alias{DEEButil}
\alias{DEEButil-package}
\title{DEEButil: Utilities for the DEEB}
\description{
Utilities for the DEEB.
}
\author{
\strong{Maintainer}: Christof Schötz \email{math@christof-schoetz.de} (\href{https://orcid.org/0000-0003-3528-4544}{ORCID})

}
\keyword{internal}
