% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{evaluateMonomials}
\alias{evaluateMonomials}
\title{Evaluate Monomials.}
\usage{
evaluateMonomials(x, degrees)
}
\arguments{
\item{x}{A n x d matrix. The n input vectors, each of dimension d.}

\item{degrees}{A p x d matrix. The degrees of the p output monomial for each of the d dimensions.}
}
\value{
A n x p matrix. The values of the p monomials for each of the n inputs vectors.
}
\description{
Evaluate Monomials.
}
