% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{getMonomialExponents}
\alias{getMonomialExponents}
\title{Get exponents of a degree deg polynomial in dim dimensions.}
\usage{
getMonomialExponents(dimension, degree)
}
\arguments{
\item{dimension}{A positive integer. The dimension of the space.}

\item{degree}{A nonegative integer. The degreee of the polynomial.}
}
\value{
A p x d integer matrix. The exponents for the p = (dim + deg choose deg) terms of the polynomial.
}
\description{
Get exponents of a degree deg polynomial in dim dimensions.
}
