% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{whichMinDist}
\alias{whichMinDist}
\title{Get the index of the closest row.}
\usage{
whichMinDist(target, query)
}
\arguments{
\item{target}{A n x p matrix.}

\item{query}{A vector of length p.}
}
\value{
The smallest index of a row of target that has the smallest
Euclidean distance to query.
}
\description{
Get the index of the closest row.
}
